This is a large commit that starts to glue a few moving pieces (configuration files, powershell scripts, mcollective and sshd).
The code will be broken until oo-scripts are updated to load modules from the correct location.

Moved oo-* scripts to the VS Powershell Project.
Added post-build events for the Uhuru.OpenShift.Node assembly that copies powershell scripts and the mcollective agent to the target bin directory.

Added script to configure mcollective facts.
Moved mcollective setup script to the powershell project. Also changed the script to download the msi from the internet.

Modified the configuration class - added a NodeConfig class in the Uhuru.OpenShift.Node assembly.
The node configuration is now used properly instead of hardcoded values. It is also used for fact generation.

Added an empty console application, oo-trap-user that will be used as a shell for users in sshd.

Change-Id: Id240f5b9a1c385411b01c302965ee8933389f623